Move the client connector creation logic to init action impl.
